msp/mgmt: move TestLocalMSP into mgmt_test
