+++ msp/mgmt/mgmt_test.go
@@ -146,6 +146,21 @@ func LoadMSPSetupForTesting() (bccsp.BCCSP, error) {
 	return cryptoProvider, nil
 }
 
+func TestLocalMSP(t *testing.T) {
+	mspDir := configtest.GetDevMspDir()
+	err := LoadLocalMsp(mspDir, nil, "SampleOrg")
+	if err != nil {
+		t.Fatalf("LoadLocalMsp failed, err %s", err)
+	}
+
+	cryptoProvider, err := sw.NewDefaultSecurityLevelWithKeystore(sw.NewDummyKeyStore())
+	require.NoError(t, err)
+	_, err = GetLocalMSP(cryptoProvider).GetDefaultSigningIdentity()
+	if err != nil {
+		t.Fatalf("GetDefaultSigningIdentity failed, err %s", err)
+	}
+}
+
 func TestMain(m *testing.M) {
 	mspDir := configtest.GetDevMspDir()
